changed substitution to using an alias method, not yet tested
git-svn-id: svn://scit.us/dawg/current@677 7e5b8c5c-87e2-0310-a459-8d91125b37d5

+++ src/include/dawg/residue.h
@@ -160,7 +160,7 @@ public:
 				
 		cs_decode = &mods[a*64];
 
-		_gap = strchr(cs_decode, '-')-cs_decode;
+		_gap = static_cast<unsigned int>(strchr(cs_decode, '-')-cs_decode);
 
 		if(a < AA)
 			cs_encode = &rmods[0*80];
